Update profile-3d.yml
This GitHub Actions workflow, named "GitHub-Profile-3D-Contrib," is scheduled to run daily at 18:00 UTC and can be manually triggered. It generates a 3D contribution graph for the GitHub profile using the yoshi389111/github-profile-3d-contrib action, commits the generated files with the user name "idimetrix," and pushes the changes to the repository.